mcp-onboarding/requirements.txt
flask
pymssql
python-dotenv
mcp[cli]
